Remove errant baseurl from Jekyll configuration
The "baseurl" configuration is intended to include the portion of the
site URL after the domain and the Jekyll page path*. In the case where
you don't want anything in between, the default value of "" (not /) is
correct.

We could have "" in the configuration to explicitly reiterate the
default setting, but some of our website building machinery doesn't
handle quotations. Instead, it seems reasonable to just remove the
setting and implicitly use the default.

* https://byparker.com/blog/2014/clearing-up-confusion-around-baseurl/