Remove HTMLProofer cache
The cache for HTMLProofer makes testing more efficient by only checking
for broken links that haven't been checked recently.

We don't have very many external links, and they can be checked very
quickly. So, this isn't buying much for us.

Removing this also has the benefit of no longer creating a cache
directory (tmp/.htmlproofer) in the repository. This was somewhat
problematic if using a docker container as it would be created owned by
root.