Use separate temp directory in rake test
The existing rake test created a _rakesite directory in the repository
for building the site. Instead, create it under /tmp so it doesn't
interfere with the repository.

This is especially useful in the case of using a container to run the
rake tests, since the directory is created with root ownership.