Add generated documentation for Bazel rules
This change adds another documentation type to build.sh: Bazel files.
Bazel files have their documentation generated by the stardoc tool. The
tool does not have a feature to add Jekyll headers, so the addition to
the build script extracts the header from a comment in the original .bzl
file. The source manifest lists both the build target and the original
.bzl file for each page of documentation.

The generated documentation added by this change is with a stardoc that
is modified to escape the HTML tags of docstrings, such as the "<name>"
in asylo.bzl. There is an open PR to incorporate this fix upstream.

The *_deps.bzl files are still have hand-documented rules due to an
existing bug in stardoc that prevents depending on libraries in the
@bazel_tools workspace. The installation_path argument of sgx_deps is
now part of asylo_deps, since the toolchain is independent of backend.