Add relocate_images to build.sh
This addition to build.sh interprets a markdown image directive as
alternative text and relocates the source image to a destination path.
The alternative text may have the initial description and the stated
destination values interpolated into the replacement text as
$description and $destination.

Documentation that references images may now have their dependencies
forwarded to the right website location.