Add a more general syntax for image relocation
This syntax allows more complex content to be replaced at the cost of
being more verbose: there is no inference for the image description, so
that must be explicitly written in the replacement text.

Documentation added for the syntactic form, including a path update to
the protoc-gen-docs tool in the istio tools repo.